# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "药房进销存与处方调配后端服务",
        "title": "药房管理系统 API",
        "contact": {},
        "version": "1.3.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/adverse-reactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "adverse-reactions"
                ],
                "summary": "不良反应列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "药品ID",
                        "name": "drug_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "adverse-reactions"
                ],
                "summary": "新建不良反应登记",
                "parameters": [
                    {
                        "description": "登记记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.AdverseReaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/adverse-reactions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "adverse-reactions"
                ],
                "summary": "更新不良反应登记",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "登记ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "登记记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.AdverseReaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "adverse-reactions"
                ],
                "summary": "删除不良反应登记",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "登记ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/consultations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "consultations"
                ],
                "summary": "用药咨询列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "consultations"
                ],
                "summary": "新建用药咨询",
                "parameters": [
                    {
                        "description": "咨询记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.Consultation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/consultations/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "consultations"
                ],
                "summary": "更新用药咨询",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "咨询ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "咨询记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.Consultation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "consultations"
                ],
                "summary": "删除用药咨询",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "咨询ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/medication-guidances": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "medication-guidances"
                ],
                "summary": "用药指导列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "medication-guidances"
                ],
                "summary": "新建用药指导",
                "parameters": [
                    {
                        "description": "指导记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.MedicationGuidance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/medication-guidances/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "medication-guidances"
                ],
                "summary": "更新用药指导",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "指导ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "指导记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.MedicationGuidance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "medication-guidances"
                ],
                "summary": "删除用药指导",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "指导ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handler.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.AdverseReaction": {
            "type": "object",
            "properties": {
                "batch_no": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "drug_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "outcome": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "integer"
                },
                "patient_name": {
                    "type": "string"
                },
                "reaction_desc": {
                    "type": "string"
                },
                "report_date": {
                    "type": "string"
                },
                "reporter": {
                    "type": "string"
                },
                "severity": {
                    "description": "1轻 2中 3重",
                    "type": "integer"
                }
            }
        },
...
        "yaofang_internal_model.Consultation": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "consultant": {
                    "type": "string"
                },
                "consulted_at": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "drug_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "patient_id": {
                    "type": "integer"
                },
                "patient_name": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.MedicationGuidance": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "drug_id": {
                    "type": "integer"
                },
                "guided_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "patient_name": {
                    "type": "string"
                },
                "pharmacist": {
                    "type": "string"
                },
                "prescription_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}